requirements.txt
Flask==3.0.0
sqlitecloud==0.2.0
flask-cors==4.0.0
flask-jwt-extended==4.5.0
pandas==2.2.2
openpyxl==3.1.2
pyjwt==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.1
pywebview